Centralize lapic construction
