+++ openhcl/virt_mshv_vtl/src/processor/mshv/arm64.rs
@@ -115,7 +115,7 @@ impl BackingPrivate for HypervisorBackedArm64 {
 
     fn new(params: BackingParams<'_, '_, Self>) -> Result<Self, Error> {
         vp::Registers::at_reset(&params.partition.caps, params.vp_info);
-        let _ = (params.runner, &params.backing_shared);
+        let _ = (params.runner, &params.backing_shared, params.lapics);
         Ok(Self {
             deliverability_notifications: Default::default(),
             next_deliverability_notifications: Default::default(),
