+++ openhcl/virt_mshv_vtl/src/processor/mshv/mod.rs
@@ -3,7 +3,6 @@
 
 //! Processor support for Microsoft hypervisor-backed partitions.
 
-pub mod apic;
 pub mod arm64;
 mod tlb_lock;
 pub mod x64;
